# services/control-api/openapi.yaml
openapi: 3.1.0
info:
  title: Control API
  version: 0.1.0
  description: >
    Control-plane API for managing workspaces, pipelines, connectors, schemas, and observability.
    Provides resource CRUD and Kafka admin operations for the Event Streaming Platform.
  contact:
    name: Platform Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://localhost:8080/api
    description: API base path
security:
  - BearerAuth: []
paths:
...
  /api/pipelines:
    get:
      summary: List all pipelines
      operationId: listPipelines
      tags:
        - Pipelines
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PageQuery
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PerPageQuery
      responses:
        "200":
          description: List of pipelines
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/ListResponse
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/Pipeline
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

    post:
      summary: Create a new pipeline
      operationId: createPipeline
      tags:
        - Pipelines
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/CreatePipelineRequest
      responses:
        "201":
          description: Pipeline created successfully
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/Pipeline
        "400":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/ValidationError
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError
  /api/pipelines/{pipelineId}:
    get:
      summary: Get a single pipeline by ID
      operationId: getPipeline
      tags:
        - Pipelines
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PipelineIdParam
      responses:
        "200":
          description: Pipeline details
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/Pipeline
        "400":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/BadRequest
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "404":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/NotFound
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

    put:
      summary: Update a pipeline
      operationId: updatePipeline
      tags:
        - Pipelines
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PipelineIdParam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/UpdatePipelineRequest
      responses:
        "200":
          description: Pipeline updated successfully
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/Pipeline
        "400":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/ValidationError
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "404":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/NotFound
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError
...
components:
  securitySchemes:
    BearerAuth:
      $ref: ../../packages/openapi-components/control-api.yaml#/components/securitySchemes/BearerAuth
...